requirements.txt
streamlit
pandas
numpy
scipy
plotly
openpyxl
python-docx
Pillow
statsmodels
scikit-posthocs
